Set margins for page data.frame (adding body width, height, and asp)

Set margins for page data.frame (adding body width, height, and asp)

Usage

set_page_margin(
  page = NULL,
  margins,
  unit = "in",
  cols = c("width", "height"),
  ...
)

Arguments

page

A character vector with a page size name or a data.frame. Passed to x parameter of as_page().

margins

Passed to get_margin() with unit value.

unit

Unit used for the margin. If margin is a unit object, unit is ignored. If page uses different units, the margins are converted into the page units for consistency.

cols

Column names to use for width and height columns. Defaults to c("width", "height"). Must be length 2 and the first value is always used as as the width name and the second as the height.

...

Passed to as_page() with page and cols.

Value

A data.frame with the page dimensions and additional columns for body dimensions, body aspect ratio, and margins.

See also